Fix the client traffic lookup by name
When the used traffic is collected, the database is searched
for the client email by partial look up. If there are similar
names with a pre or post-fix sometimes they will be mixed up
and the usage doesn't get updated. Fixed it by adding
quotation mark around the client email.