Bot: Send payment details automatically on purchase request
Payment details message can be configured via the settings.
The message will be sent to the user and wait for the receipt
before the order is completed and shown on the panel.

So the only step left in the new order process is for the admin
to approve the order on the panel.